requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.5
python-dotenv==1.0.1
pydantic-settings==2.8.1
PyJWT==2.10.1
passlib==1.7.4
redis==5.2.1
httpx==0.28.1
